CL - fix pengguna

+++ app/Http/Controllers/PenggunaController.php
@@ -301,12 +301,18 @@ class PenggunaController extends Controller
         $pro_peserta = MohonPenilaian::where('no_ic', $ic)->first();
         if ($pro_peserta != null) {
             $pro_peserta->delete();
-        }
-        $pro_tempat_tugas = Tugas::where('ID_PESERTA', $pro_peserta->ID_PESERTA)->first();
+	}
+
+	$pro_peserta_2 = Permohanan::where('NO_KAD_PENGENALAN',$ic)->first();
+	if ($pro_peserta_2 != null){
+		$pro_peserta_2->delete();
+	}
+
+        $pro_tempat_tugas = Tugas::where('ID_PESERTA', $pro_peserta_2->ID_PESERTA)->first();
         if ($pro_tempat_tugas != null) {
             $pro_tempat_tugas->delete();
         }
-        $pro_perkhidmatan = Perkhidmatan::where('ID_PESERTA', $pro_peserta->ID_PESERTA)->first();
+        $pro_perkhidmatan = Perkhidmatan::where('ID_PESERTA', $pro_peserta_2->ID_PESERTA)->first();
         if ($pro_perkhidmatan != null) {
             $pro_perkhidmatan->delete();
         }
@@ -323,8 +329,12 @@ class PenggunaController extends Controller
             }
         }
 
-        $kemahiran = Bankjawapancalon::where('id_calon', $ic)->get();
-        $kemahiran->delete();
+	$kemahiran = Bankjawapancalon::where('ic_calon', $ic)->get();
+	if($kemahiran != null){
+		foreach($kemahiran as $kemahiran){
+			$kemahiran->delete();
+		}
+	}
 
         $user->delete();
         return redirect('/pengurusanpengguna')->with('success', 'Berjaya dihapus!');
